Renaming of RecordClasses
