+++ src/nuis/measurement/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(measurement SHARED HEPDataLoader.cxx  Projection.cxx  SimpleStatistics.cxx)
+add_library(measurement SHARED HEPDataRecord.cxx  Projection.cxx Statistics.cxx)
 
 target_link_libraries(measurement PUBLIC nuis_options)
